Sidebar favourites (Videos/Music) test
Reusing Kicker.FavoritesModel worked well. Unfortunately, the music/video folder icons at 22px aren't symbolic, and there doesn't seem to be a way to get the iconName, so we'll need to map a hardcoded list of paths to specific icons. Another problem is that we need to dynamically pass the home directory path instead of serializing it (probably).

In summary, it's probably easier to make a hardcoded list of folders + icons.